requirements.txt
agent-framework-azure-ai>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
azure-identity>=1.14.0
